refactor: update data structure in API responses and adjust related components
